Move Pre to own component
